hw4/requirements.txt
numpy
sklearn
Pillow
matplotlib
word2vec
adjustText